+++ tests/invalid_ast_samples/567b705d091d5a43efcaf38c9980921e1ee819e1c33484d90308b96d57acb896.py
@@ -1,0 +1,34 @@
+from ast import arguments
+from ast import FunctionDef
+from ast import Global
+from ast import Module
+from ast import Nonlocal
+
+tree = Module(
+    body=[
+        FunctionDef(
+            name="name_1",
+            args=arguments(),
+            body=[
+                Global(names=["name_2"]),
+                FunctionDef(
+                    name="name_2", args=arguments(), body=[Nonlocal(names=["name_2"])]
+                ),
+            ],
+        )
+    ]
+)
+
+# version: 3.13.0b1
+# seed = 2636981
+#
+# Source:
+# def name_1():
+#     global name_2
+#
+#     def name_2():
+#         nonlocal name_2
+#
+#
+# Error:
+#     SyntaxError("no binding for nonlocal 'name_2' found")