+++ tests/invalid_ast_samples/d8fe8ca9a8f439de6b8552a8b39ec22a7efca65bcadcbb9945cc01775b6d5af8.py
@@ -1,0 +1,28 @@
+from ast import Load
+from ast import Module
+from ast import Name
+from ast import NamedExpr
+from ast import Store
+from ast import TypeAlias
+
+tree = Module(
+    body=[
+        TypeAlias(
+            name=Name(id="name_0", ctx=Store()),
+            value=NamedExpr(
+                target=Name(id="name_2", ctx=Store()),
+                value=Name(id="name_0", ctx=Load()),
+            ),
+        )
+    ]
+)
+
+# version: 3.13.0b1
+# seed = 1933771
+#
+# Source:
+# type name_0 = (name_2 := name_0)
+#
+#
+# Error:
+#     SyntaxError('named expression cannot be used within a type alias')