+++ tests/valid_source_samples/5c2f9e9ccf5b6b88b118bdcdb0497214f9884f839b82d9a75da2577090f4ec47.py
@@ -1,0 +1,1 @@
+global name_0, name_0